Add files via upload
This is an implementation of an Algorithm to detect, classify and localize objects and give the real-world coordinates of the object. This project uses the COCO dataset and state-of-the art SSD object Detection algorithm to detect these objects. This algorithm takes real-time input from a Microsoft Kinect camera, which also has a depth sensor to measure the depth of the objects. The value of the depth will further be used by this algorithm to extract the real world coordinates.